Move smoke-runner to vrt-markup; extract smoke types
The a11y-driven smoke test runner is a vrt-markup concern (it
inspects accessibility trees and drives interactions), not a CLI
detail. Move it next to the rest of the inspect/* modules.

- `src/cli/commands/smoke-runner.ts` → `packages/vrt-markup/src/inspect/smoke-runner.ts`
- New `packages/vrt-markup/src/inspect/smoke-types.ts` owns the
  SmokeTestRequest/Response/Action/Error + A11ySnapshot/A11yNodeCompact
  types (previously co-located with HTTP API types in src/api/api-types.ts).
- `src/api/api-types.ts` re-exports the smoke types so HTTP consumers
  keep their existing import path.
- `src/api/api-server.ts` imports `runSmokeTest` from the new package
  path.
- router.ts + Taskfile.pkl point at the new file location.

Note: tried to extract `src/api/` to a vrt-api package too, but
api-server lazy-imports `src/experiments/css-challenge/css-challenge-core`
(for `capturePageState` + `diffComputedStyles` on the /compare
endpoint). That cross-cuts the experiments tree and would force a much
larger move; left api/ in src/ for now.
